requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
roboflow==1.0.9
opencv-python==4.5.3.56
gunicorn==20.1.0